And more fixes, total as decimal

+++ SurveyScorer/Entities/Response/ScoreCard.cs
@@ -5,7 +5,7 @@ namespace SurveyScorer.Entities.Response
 {
     public class ScoreCard
     {
-        public int Aggregate { get; set; }
+        public decimal Aggregate { get; set; }
         public ResultColor ResultColor { get; set; }
         public IList<MetaResponse> Metadata { get; set; }
         public IList<CategoryScore> ScoreItems { get; set; }
